Entity Framework User-Character Relationship

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
@@ -13,26 +14,32 @@ namespace dotnet_rpg.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _dbContext;
+        private readonly IHttpContextAccessor _contextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext dbContext)
+        public CharacterService(IMapper mapper, DataContext dbContext, IHttpContextAccessor contextAccessor)
         {
             _mapper = mapper;
             _dbContext = dbContext;
+            _contextAccessor = contextAccessor;
         }
         private static List<Character> characters = new List<Character>{
             new Character(),
             new Character{ Id = 1,Name = "John",Class = RpgClass.Healer}
         };
 
+        private int GetUserId() => int.Parse(_contextAccessor.HttpContext.User
+            .FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto character)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character chara = _mapper.Map<Character>(character);
+            chara.User = await _dbContext.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             _dbContext.Characters.Add(chara);
             await _dbContext.SaveChangesAsync();
-            serviceResponse.Data = await _dbContext.Characters.Select(x =>
-                    _mapper.Map<GetCharacterDto>(x))
+            serviceResponse.Data = await _dbContext.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(x => _mapper.Map<GetCharacterDto>(x))
                     .ToListAsync();
             return serviceResponse;
         }
@@ -40,15 +47,16 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> Get(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _dbContext.Characters.FirstOrDefaultAsync(x => x.Id == id);
+            var dbCharacter = await _dbContext.Characters
+                .FirstOrDefaultAsync(x => x.Id == id && x.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAll(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAll()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _dbContext.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            var dbCharacters = await _dbContext.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbCharacters.Select(x => _mapper.Map<GetCharacterDto>(x)).ToList();
             return serviceResponse;
         }
@@ -60,17 +68,26 @@ namespace dotnet_rpg.Services.CharacterService
             try
             {
                 var character = await _dbContext.Characters
+                    .Include(c => c.User)
                     .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
-                character.Name = updatedCharacter.Name;
-                character.HitPoints = updatedCharacter.HitPoints;
-                character.Strength = updatedCharacter.Strength;
-                character.Defence = updatedCharacter.Defence;
-                character.Intelligence = updatedCharacter.Intelligence;
-                character.Class = updatedCharacter.Class;
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updatedCharacter.Name;
+                    character.HitPoints = updatedCharacter.HitPoints;
+                    character.Strength = updatedCharacter.Strength;
+                    character.Defence = updatedCharacter.Defence;
+                    character.Intelligence = updatedCharacter.Intelligence;
+                    character.Class = updatedCharacter.Class;
+                    await _dbContext.SaveChangesAsync();
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
 
-                await _dbContext.SaveChangesAsync();
-                response.Data = _mapper.Map<GetCharacterDto>(character);
             }
             catch (Exception ex)
             {
@@ -85,12 +102,22 @@ namespace dotnet_rpg.Services.CharacterService
             ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _dbContext.Characters.FirstAsync(c => c.Id == id);
-                _dbContext.Characters.Remove(character);
-                await _dbContext.SaveChangesAsync();
-                response.Data = await _dbContext.Characters.Select(x =>
-                    _mapper.Map<GetCharacterDto>(x))
-                    .ToListAsync();
+                Character character = await _dbContext.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _dbContext.Characters.Remove(character);
+                    await _dbContext.SaveChangesAsync();
+                    response.Data = await _dbContext.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                        .Select(x => _mapper.Map<GetCharacterDto>(x))
+                        .ToListAsync();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (Exception ex)
             {
